Fixed a bug that caused the script to crush if a word's definition had no usage example(s)
If one definition of a word didn't have a usage example on the website, the script checked for it as an element of a list and throwed an error because it didn't find one. It was true, for example, for a word "дорога".

Now it appends a string "Отсутствует пример употребления" (en. "Lacks usage example") that is being treated as a from-dictionary-fetched example later in the script.